feat: add left sidebar
